Blazor project created with working Transactions visualizer, Login, and Register pages. Authentication implemented for both backend and frontend. Api improved, full transactions logic implemented

+++ PersonalFinanceApp.Api/PersonalFinanceApp.Api/Controllers/TransactionsController.cs
@@ -1,5 +1,7 @@
 ﻿using BaseLibrary.DTOs;
+using BaseLibrary.Entities;
 using BaseLibrary.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using PersonalFinanceApp.Api.Repositories.Contracts;
 
@@ -9,6 +11,7 @@ namespace PersonalFinanceApp.Api.Controllers
 {
     [Route("api/transactions")]
     [ApiController]
+    [Authorize]
     public class TransactionsController : ControllerBase
     {
         private readonly ITransactionRepository _transactionRepository;
@@ -17,11 +20,14 @@ namespace PersonalFinanceApp.Api.Controllers
         {
             _transactionRepository = transactionRepository;
         }
+
+        #region api/transactions
+
         // GET: api/transactions
         [HttpGet]
         public async Task<ActionResult<IEnumerable<TransactionDTO>>> GetTransactions()
         {
-            var transactions = await _transactionRepository.GetTransactions();
+            var transactions = await _transactionRepository.GetTransactions(User.GetUserId());
 
             if (transactions == null)
                 return NoContent();
@@ -34,13 +40,13 @@ namespace PersonalFinanceApp.Api.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<TransactionDTO>> GetTransaction(long id)
         {
-            var transaction = await _transactionRepository.GetTransaction(id);
+            var transaction = await _transactionRepository.GetTransaction(User.GetUserId(), id);
             if (transaction == null)
                 return NotFound();
 
             var transactionDTO = transaction.ConvertToDTO();
             return Ok(transactionDTO);
-        }        
+        }
 
         // POST api/transactions
         [HttpPost]
@@ -48,7 +54,10 @@ namespace PersonalFinanceApp.Api.Controllers
         {
             if (transactionToAdd == null)
                 throw new BadHttpRequestException(new ArgumentNullException().Message);
-            var transaction = await _transactionRepository.AddTransaction(transactionToAdd.ConvertToEntity());
+            var transactionEntity = transactionToAdd.ConvertToEntity();
+            transactionEntity.UserId = User.GetUserId();
+            transactionEntity.Update();
+            var transaction = await _transactionRepository.AddTransaction(transactionEntity);
             return CreatedAtAction(
                 nameof(GetTransaction),
                 new { id = transaction.Id },
@@ -57,28 +66,62 @@ namespace PersonalFinanceApp.Api.Controllers
 
         // PUT api/transactions/5
         [HttpPut("{id}")]
-        public async Task<ActionResult<TransactionDTO>> UpdateTransaction(long id, [FromBody] TransactionDTO transactionToUpdate)
+        public async Task<ActionResult<TransactionDTO>> UpdateTransaction(long id, [FromBody] TransactionDTO transaction)
         {
-            if (transactionToUpdate == null)
-                throw new BadHttpRequestException(new ArgumentNullException().Message);
-            if (id != transactionToUpdate.Id)            
+            if (transaction == null)
+                throw new BadHttpRequestException(new ArgumentNullException().Message);            
+
+            if (id != transaction.Id)
                 return BadRequest();
 
-            var transaction = await _transactionRepository.UpdateTransaction(id, transactionToUpdate.ConvertToEntity());
-            if (transaction == null)            
-                return NotFound();            
-            
-            return Ok(transaction.ConvertToDTO());
+            bool ownsTransaction = await _transactionRepository.UserOwnsTransaction(User.GetUserId(), id);
+
+            if (!ownsTransaction)
+                return BadRequest();
+
+            var transactionEntity = transaction.ConvertToEntity();
+            transactionEntity.UserId = User.GetUserId();
+            transactionEntity.Update();
+            var transactionF = await _transactionRepository.UpdateTransaction(id, transactionEntity);
+            if (transactionF == null)
+                return NotFound();
+
+            return Ok(transactionF.ConvertToDTO());
         }
 
         // DELETE api/transactions/5
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteTransaction(long id)
         {
+            bool ownsTransaction = await _transactionRepository.UserOwnsTransaction(User.GetUserId(), id);
+
+            if (!ownsTransaction)
+                return BadRequest();
+
             var transaction = await _transactionRepository.DeleteTransaction(id);
-            if(transaction==null)
+            if (transaction == null)
                 return NotFound();
             return NoContent();
         }
+
+        #endregion
+
+        #region api/transactions/[misc]
+
+
+        // GET: api/transactions
+        [AllowAnonymous]
+        [HttpGet("types")]
+        public async Task<ActionResult<IEnumerable<TransactionType>>> GetTransactionTypes()
+        {
+            var types = await _transactionRepository.GetTransactionTypes();
+
+            if (types == null)
+                return NoContent();
+            
+            return Ok(types);
+        }       
+
+        #endregion
     }
 }
